change types

+++ types.go
@@ -10,7 +10,7 @@ import (
 type APIResponse struct {
 	Status  string          `json:"status"`
 	Data    json.RawMessage `json:"data"`
-	RetCode int64           `json:"retcode"`
+	RetCode int             `json:"retcode"`
 }
 
 // Update is an update response, from GetUpdates.
@@ -18,16 +18,16 @@ type Update struct {
 	PostType      string      `json:"post_type"`
 	MessageType   string      `json:"message_type"`
 	SubType       string      `json:"sub_type"`
-	MessageID     int         `json:"message_id"`
-	GroupID       int         `json:"group_id"`
-	DiscussID     int         `json:"discuss_id"`
-	UserID        int         `json:"user_id"`
+	MessageID     int64       `json:"message_id"`
+	GroupID       int64       `json:"group_id"`
+	DiscussID     int64       `json:"discuss_id"`
+	UserID        int64       `json:"user_id"`
 	Font          int         `json:"font"`
 	RawMessage    interface{} `json:"message"`
 	AnonymousName string      `json:"anonymous"`
 	AnonymousFlag string      `json:"anonymous_flag"` // Anonymous ID
 	Event         string      `json:"event"`
-	OperatorID    int         `json:"operator_id"`
+	OperatorID    int64       `json:"operator_id"`
 	File          *File       `json:"file"`
 	Flag          string      `json:"flag"`
 	Text          string      `json:"-"` // Known as "message", in a message or request
@@ -46,7 +46,7 @@ type File struct {
 
 // User is a user on QQ.
 type User struct {
-	ID       int    `json:"user_id"`
+	ID       int64  `json:"user_id"`
 	NickName string `json:"nickname"`
 	Sex      string `json:"sex"` // "male"、"female"、"unknown"
 	Age      int    `json:"age"`
@@ -109,7 +109,7 @@ func (c Chat) IsDiscuss() bool {
 // almost anything.
 type Message struct {
 	*cqcode.Message `json:"message"`
-	MessageID       int    `json:"message_id"`
+	MessageID       int64  `json:"message_id"`
 	From            *User  `json:"from"`
 	Chat            *Chat  `json:"chat"`
 	Text            string `json:"text"`
